requirements.txt
asn1crypto==0.24.0
backports.ssl-match-hostname==3.5.0.1
cached-property==1.3.1
certifi==2018.1.18
chardet==3.0.4
click==7.1.1
crcmod==1.7
cryptography==2.1.4
dnspython==1.16.0
docker==2.5.1
docker-compose==1.17.1
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
enum34==1.1.10
eventlet==0.25.0
Flask==1.1.1
funcsigs==1.0.2
functools32==3.2.3.post2
greenlet==0.4.15
gunicorn==19.10.0
gyp==0.1
html==1.16
idna==2.6
ipaddress==1.0.17
itsdangerous==1.1.0
Jinja2==2.11.1
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
libsass==0.19.2
MarkupSafe==1.1.1
mock==2.0.0
monotonic==1.5
pbr==3.1.1
powerline-status==2.6
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pygobject==3.26.1
pyOpenSSL==17.5.0
python-engineio==3.12.1
python-socketio==4.3.1
pyxdg==0.25
PyYAML==3.12
requests==2.18.4
SecretStorage==2.3.1
six==1.14.0
texttable==0.9.1
urllib3==1.22
websocket-client==0.44.0
Werkzeug==1.0.1
